requirements.txt
robin-stocks>=3.4.0
pandas>=2.0.0
numpy>=1.24.0
pyotp>=2.9.0
python-dotenv>=1.0.0
loguru>=0.7.0
schedule>=1.2.0
anthropic>=0.25.0
sendgrid>=6.10.0
rich>=13.0.0
pytz>=2023.3
requests>=2.31.0
colorama>=0.4.6
tabulate>=0.9.0
